// currapay-frontend/app/profile/types.ts
export type userData = {
    age: number;
    browserUsed: string;
    city: string;
    country: string;
    createdAt: string;
    deviceUsed: string;
    emailAddress: string;
    gender: string;
    id: number;
    nationality: string;
    occupation: string;
    password: string;
    updatedAt: string;
  }